+++ src/commands/find_expenses.rs
@@ -1,11 +1,20 @@
-use crate::{errors::CommandError, expense::Expense, trace_command};
+use crate::{
+    consts::{DEBUG_START, DEBUG_SUCCESS},
+    errors::CommandError,
+    expense::Expense,
+    trace_command,
+};
 use macro_rules_attribute::apply;
 use teloxide::prelude::*;
 use tracing::Level;
 
 #[apply(trace_command)]
 pub async fn find_expenses(msg: &Message, description: &str) -> Result<String, CommandError> {
-    tracing::debug!("START");
+    tracing::debug!(DEBUG_START);
+    if description.is_empty() {
+        return Err(CommandError::EmptyInput);
+    }
+
     let list_res = Expense::db_select_by_descr(msg.chat.id, description.to_owned()).await;
     match list_res {
         Ok(expenses) => {
@@ -18,7 +27,7 @@ pub async fn find_expenses(msg: &Message, description: &str) -> Result<String, C
                     .collect::<Vec<_>>()
                     .join("\n")
             };
-            tracing::debug!("SUCCESS");
+            tracing::debug!(DEBUG_SUCCESS);
             Ok(reply)
         }
         Err(err) => {