+++ src/commands/mod.rs
@@ -8,6 +8,8 @@ mod help;
 mod help_message;
 mod list_expenses;
 mod list_travelers;
+mod show_balance;
+mod show_balances;
 mod transfer;
 mod unknown_command;
 
@@ -21,5 +23,7 @@ pub use help::help;
 pub use help_message::HelpMessage;
 pub use list_expenses::list_expenses;
 pub use list_travelers::list_travelers;
+pub use show_balance::show_balance;
+pub use show_balances::show_balances;
 pub use transfer::transfer;
 pub use unknown_command::unknown_command;