+++ src/views/balance.rs
@@ -5,43 +5,26 @@ use surrealdb::RecordId;
 use teloxide::types::ChatId;
 use travel_rs_derive::Table;
 
+const BALANCES_VIEW: &str = "v_balances";
+
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Table)]
-pub struct Owes {
+pub struct Balance {
     pub id: RecordId,
-    pub amount: Decimal,
-    pub r#in: RecordId,
-    pub out: RecordId,
+    pub debt: Decimal,
+    pub debtor_name: Name,
+    pub creditor_name: Name,
+    pub chat: RecordId,
 }
 
-impl Owes {
-    pub async fn db_relate(
-        amount: Decimal,
-        debitor: RecordId,
-        creditor: RecordId,
-    ) -> Result<Option<Self>, surrealdb::Error> {
-        let db = db().await;
-        db.query(format!(
-            "RELATE ${IN}->{TABLE}->${OUT}
-            SET {AMOUNT} = <decimal> ${AMOUNT}",
-        ))
-        .bind((IN, debitor))
-        .bind((OUT, creditor))
-        .bind((AMOUNT, amount))
-        .await
-        .and_then(|mut response| response.take::<Option<Self>>(0))
-    }
-
+impl Balance {
     pub async fn db_select(chat_id: ChatId) -> Result<Vec<Self>, surrealdb::Error> {
-        use crate::{
-            chat::{ID as CHAT_ID, TABLE as CHAT_TB},
-            traveler::CHAT,
-        };
+        use crate::chat::{ID as CHAT_ID, TABLE as CHAT_TB};
 
         let db = db().await;
         db.query(format!(
             "SELECT *
-            FROM {TABLE}
-            WHERE {IN}.{CHAT}.{CHAT_ID} = ${CHAT_ID}",
+            FROM {BALANCES_VIEW}
+            WHERE {CHAT} = ${CHAT_ID}",
         ))
         .bind((CHAT_ID, RecordId::from_table_key(CHAT_TB, chat_id.0)))
         .await
@@ -54,16 +37,16 @@ impl Owes {
     ) -> Result<Vec<Self>, surrealdb::Error> {
         use crate::{
             chat::{ID as CHAT_ID, TABLE as CHAT_TB},
-            traveler::{CHAT, NAME},
+            traveler::NAME,
         };
 
         let db = db().await;
         db.query(format!(
             "SELECT *
-            FROM {TABLE}
+            FROM {BALANCES_VIEW}
             WHERE
-                {IN}.{CHAT}.{CHAT_ID} = ${CHAT_ID}
-                && ({IN}.{NAME} = ${NAME} || {OUT}.{NAME} = ${NAME})",
+                {CHAT} = ${CHAT_ID}
+                && ({DEBTOR_NAME} = ${NAME} || {CREDITOR_NAME} = ${NAME})",
         ))
         .bind((CHAT_ID, RecordId::from_table_key(CHAT_TB, chat_id.0)))
         .bind((NAME, name))